requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.9
pydantic==2.10.4
pillow==10.4.0
numpy==2.2.1
requests==2.32.3
python-dotenv==1.0.1
ffmpeg-python==0.2.0
